+++ ReleaseNotes.md
@@ -24,7 +24,9 @@
   For simplicity, I decided to avoid the trouble of maintaining two different types of software license and just use MIT license for all the source code contributed on top of the Microsoft sample language service provider extension, which is already licensed under MIT. The LICENSE file and the license headers have been updated to reflect the change.
 
 - New feature:
+
   - Add a setting to allow applyting the color token detection and previewing feature on any language except for css and less.
+  - Add the "Go to definition" support for referenced css variables. If json documents with color tokens are open and there is a css/less document referencing them as css variables, applying "Go to definition" on the line of referenced variable will bring the editor to the location in the json documents where the color tokens are defined.
 
 - Known issues: